hello world: use liteuart console
Really this should be a build flag that can alternate between them.
We don't really want to use syscon because a bitstream could have both
built in. The other option would be to write both UARTs, but then we
need to worry about a disconnected USB cable blocking forever on TX.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>